Full width slide down syllabus chat for the resource drawer (#2063)
* Remove columned chatbot. Refactor

* Full width / slide down syllabus bot

* Update Posthog replay vars

* Refactor chat with entry screen for reuse

* Update test and clean up

* Fix enum import

* Test updates

* Prevent tab to masked content when the chat is open (accessibility)

* Limit scroll to content area when chat enabled

* Remove scrollbar-gutter not needed

* Rewind limit scroll to content area when chat enabled

This reverts commit 823d9555a86d01c07e47ee143c69406a8557a036.

* avoid layout shift, show full scrollbar

* Make chat inert when closed. Fix pointer events stopped at opener container

---------
