+++ frontends/main/src/page-components/LearningResourceExpanded/InfoSection.test.tsx
