+++ frontends/main/src/page-components/LearningResourceExpanded/VideoFrame.tsx
